requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.0
django-froala-editor==4.0.17
idna==3.4
Pillow==9.4.0
pyparsing==3.0.9
pyrfc6266==1.0.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
